+++ src/index.ts
@@ -53,6 +53,7 @@ export type {
   PivotQueryBuilder,
 } from "./model/Model.js";
 export { ModelNotFoundError } from "./model/ModelNotFoundError.js";
+export { MassAssignmentError } from "./model/MassAssignmentError.js";
 export { Observer, ObserverRegistry, type ObserverContract } from "./model/Observer.js";
 export { MorphMap } from "./model/MorphMap.js";
 export { MorphTo, MorphOne, MorphMany, MorphToMany } from "./model/MorphRelations.js";